requirements.txt
discord.py
requests.py
json.py
dotenv.py
datetime.py